Updated some libraries.

+++ mopub/libproj/mopub/MoPubSDK/Internal/Common/MPAdServerCommunicator.m
@@ -84,7 +84,7 @@ const NSTimeInterval kRequestTimeoutInterval = 10.0;
 - (void)connection:(NSURLConnection *)connection didReceiveResponse:(NSURLResponse *)response
 {
     if ([response respondsToSelector:@selector(statusCode)]) {
-        int statusCode = [(NSHTTPURLResponse *)response statusCode];
+        NSInteger statusCode = [(NSHTTPURLResponse *)response statusCode];
         if (statusCode >= 400) {
             [connection cancel];
             self.loading = NO;
